third_party/blink/web_tests/http/tests/security/local-iFrame-from-remote-expected.txt

CONSOLE ERROR: Not allowed to load local resource: localPage.html
This test is to see if a remote file can include a local page in an iFrame.
Currently this test cannot be run manually on Windows because we do not have a function like pathToLocalResource() outside of DRT.



Test Passed. Local page not remotely loaded into iFrame.
